--- es6/marshal.js
@@ -1,4 +1,0 @@
-import dispatch from 'dispatch-recursive';
-import {marshalCommandList} from './lib/commands';
-
-export default dispatch(...marshalCommandList);
