--- es6/unmarshal.js
@@ -1,4 +1,0 @@
-import dispatch from 'dispatch-recursive';
-import {unmarshalCommandList} from './lib/commands';
-
-export default dispatch(...unmarshalCommandList);
