+++ lib/ensureItemIsObject.js
@@ -1,12 +1,4 @@
-'use strict';
-
-var _Object$defineProperty = require('babel-runtime/core-js/object/define-property')['default'];
-
-_Object$defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _isPlainObject = require('lodash');
+var _ = require('lodash');
 
 /**
  * Function decorator, validates item is a javascript object.
@@ -15,13 +7,12 @@ var _isPlainObject = require('lodash');
  * @returns {Function}
  */
 function ensureItemIsObject(fn) {
-  return function (item) {
-    if (_isPlainObject.isPlainObject(item)) {
+  return function(item) {
+    if (_.isPlainObject(item)) {
       return fn(item);
     }
     throw new TypeError('Item must be plain object literal');
   };
 }
 
-exports['default'] = ensureItemIsObject;
-module.exports = exports['default'];
+module.exports = ensureItemIsObject;
