+++ lib/unexpectedItemHandler.js
@@ -1,0 +1,5 @@
+function unexpectedItemHandler() {
+  throw new TypeError('Encountered unexpected item');
+}
+
+module.exports = unexpectedItemHandler;
