+++ marshalItem.js
@@ -1,20 +1,5 @@
-'use strict';
-
-var _Object$defineProperty = require('babel-runtime/core-js/object/define-property')['default'];
-
-var _interopRequireDefault = require('babel-runtime/helpers/interop-require-default')['default'];
-
-_Object$defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _ensureItemIsObject = require('./lib/ensureItemIsObject');
-
-var _ensureItemIsObject2 = _interopRequireDefault(_ensureItemIsObject);
-
-var _marshal = require('./marshal');
-
-var _marshal2 = _interopRequireDefault(_marshal);
+var ensureItemIsObject = require('./lib/ensureItemIsObject');
+var marshal = require('./marshal');
 
 /**
  * Translates a javascript object into a format ready for DynamoDb.
@@ -23,8 +8,7 @@ var _marshal2 = _interopRequireDefault(_marshal);
  * @returns {Object} The marshaled dynamoDb item.
  */
 function marshalItem(item) {
-  return _marshal2['default'](item).M;
+  return marshal(item).M;
 }
 
-exports['default'] = _ensureItemIsObject2['default'](marshalItem);
-module.exports = exports['default'];
+module.exports = ensureItemIsObject(marshalItem);
