+++ marshalJson.js
@@ -1,16 +1,4 @@
-'use strict';
-
-var _Object$defineProperty = require('babel-runtime/core-js/object/define-property')['default'];
-
-var _interopRequireDefault = require('babel-runtime/helpers/interop-require-default')['default'];
-
-_Object$defineProperty(exports, '__esModule', {
-  value: true
-});
-
-var _marshalItem = require('./marshalItem');
-
-var _marshalItem2 = _interopRequireDefault(_marshalItem);
+var marshalItem = require('./marshalItem');
 
 /**
  * Translates a JSON string into an object in DynamoDb format.
@@ -19,8 +7,7 @@ var _marshalItem2 = _interopRequireDefault(_marshalItem);
  * @returns {Object} The marshaled DynamoDb compliant item.
  */
 function marshalJson(json) {
-  return _marshalItem2['default'](JSON.parse(json));
+  return marshalItem(JSON.parse(json));
 }
 
-exports['default'] = marshalJson;
-module.exports = exports['default'];
+module.exports = marshalJson;
